feat(machine): raise exceptions in irregular cases

+++ MachineAPI/src/Application/Services/MachineService.cs
@@ -25,24 +25,51 @@ namespace MachineAPI.Application.Services
 
         public async Task<MachineDTO?> GetById(int id)
         {
+            if (id <= 0)
+            {
+                throw new ArgumentException("ID inválido.");
+            }
+
             Machine? machine = await _machineRepository.GetById(id);
+
+            if (machine == null)
+            {
+                throw new KeyNotFoundException($"Máquina com ID {id} não encontrada.");
+            }
+
             return machine != null ? _mapper.Map<MachineDTO>(machine) : null;
         }
 
         public async Task Add(CreateUpdateMachineDTO machineDTO)
         {
+            if (machineDTO == null)
+            {
+                throw new ArgumentNullException(nameof(machineDTO), "Nenhum dado foi recebido.");
+            }
+
             Machine machine = _mapper.Map<Machine>(machineDTO);
             await _machineRepository.Add(machine);
         }
 
         public async Task Update(int id, CreateUpdateMachineDTO machineDTO)
         {
-            Machine machine = await _machineRepository.GetById(id);
+            if (id <= 0)
+            {
+                throw new ArgumentException("ID inválido.");
+            }
+
+            if (machineDTO == null)
+            {
+                throw new ArgumentNullException(nameof(machineDTO), "Nenhum dado foi recebido.");
+            }
+
+            Machine? machine = await _machineRepository.GetById(id);
 
             if (machine == null)
             {
-                throw new KeyNotFoundException($"Machine with ID {id} not found.");
+                throw new KeyNotFoundException($"Máquina com ID {id} não encontrada.");
             }
+
             _mapper.Map(machineDTO, machine);
 
             await _machineRepository.Update(machine);
@@ -50,6 +77,18 @@ namespace MachineAPI.Application.Services
 
         public async Task Delete(int id)
         {
+            if (id <= 0)
+            {
+                throw new ArgumentException("ID inválido.");
+            }
+
+            Machine? machine = await _machineRepository.GetById(id);
+
+            if (machine == null)
+            {
+                throw new KeyNotFoundException($"Máquina com ID {id} não encontrada.");
+            }
+
             await _machineRepository.Delete(id);
         }
     }
